Check if two lattices are equal under reflection
For this commit we had to change the overall rounding precision to
1e-9, this does not make difference overall and was done in order to
sort the lattices in an unique way. Because we are looking for
symmetries in the structure, this symmetry should be robust under
small oscillations on the position of each site, therefore either we
use single or double precision for comparing then is irrelevant. We
can think of this as instead of comparing single points, we compare
small balls with radius epsilon, then we say that two lattices are
equal if theses balls overlap once we do the symmetry operation.