Include boundary on the reflections.
Each time we do a reflection we must take in account the boundary of
the problem. Because we are doing crystalline structures, we deal with
unitcells that repeat in periodically to infinity. Instead of
extending the lattice to infinity we use periodic boundary
conditions. This commit implements the `Boundary` class that handle
this boundary condition.