We need to sweep the whole range
I was mistaken in thinking that we only needed to sweep half of the
range when trying to find the reflection axis. This is not true and
there are examples of that in the problem I am working on. Of course, we
can always center the lattice symmetrical to the reflection axis. I
think this is not necessary as of now.